moved social media profile class to the profile folder
